+++ distribution/build.sh
@@ -21,9 +21,7 @@ KEYCLOAK_GIT_REPO="${KEYCLOAK_GIT_REPO:-$keycloak_git_repo}"
 KEYCLOAK_GIT_BRANCH="${KEYCLOAK_GIT_BRANCH:-$keycloak_git_branch}"
 THEME_VERSION="${THEME_VERSION:-$theme_version}"
 
-if [ "$bamboo_helm_repo_location" == "" ]; then
-  bamboo_helm_repo_location=https://kubernetes-charts.alfresco.com/incubator
-fi
+HELM_REPO_LOCATION="${HELM_REPO_LOCATION:-https://kubernetes-charts.alfresco.com/incubator}"
 
 if [ "$KEYCLOAK_GIT_REPO" != "" ]; then
   if [ "$KEYCLOAK_GIT_BRANCH" == "" ]; then
@@ -91,10 +89,10 @@ mkdir -p $DISTRIBUTION_NAME/realm
 # rendered (without installation) using helm. Note only "realm-secret.yaml" needs to be rendered as this is how the
 # realm gets passed on to keycloak when on k8s.
 #
-helm init --client-only
+helm repo add stable https://charts.helm.sh/stable
 if [ -z "$(helm repo list | grep ${HELM_REPO_NAME})" ]; then
   echo "adding helm repository"
-  helm repo add ${HELM_REPO_NAME} ${bamboo_helm_repo_location}
+  helm repo add ${HELM_REPO_NAME} "${HELM_REPO_LOCATION}"
 fi
 helm repo add codecentric https://codecentric.github.io/helm-charts
 
@@ -102,7 +100,7 @@ helm repo update
 helm dependency update ${CHART_DIR}
 helm template ${CHART_DIR} \
   -f ${CHART_DIR}/values.yaml \
-  -x templates/realm-secret.yaml \
+  --show-only templates/realm-secret.yaml \
   --set realm.alfresco.client.redirectUris='{*}' |
   grep '\"alfresco-realm.json\"' | awk '{ print $2}' |
   sed -e 's/\"$//' -e 's/^\"//' | base64 --decode | jq '.' >$DISTRIBUTION_NAME/realm/alfresco-realm.json
@@ -123,7 +121,7 @@ echo -e "set \"JAVA_OPTS=%JAVA_OPTS% -Dkeycloak.import=%~dp0..\\\realm\\\alfresc
 echo '# Alfresco realm import ' >>$DISTRIBUTION_NAME/bin/standalone.conf.ps1
 echo "\$JAVA_OPTS += \"-Dkeycloak.import=\$pwd\\\..\\\realm\\\alfresco-realm.json\"" >>$DISTRIBUTION_NAME/bin/standalone.conf.ps1
 
-# zip the confiured distro
+# zip the configured distro
 zip -rq $DISTRIBUTION_NAME.zip $DISTRIBUTION_NAME
 openssl md5 $DISTRIBUTION_NAME.zip >$DISTRIBUTION_NAME.md5