+++ distribution/tests/endpoints_ps.ps1
@@ -16,7 +16,7 @@ $COUNTER_MAX = 60
 $SLEEP_SECONDS = 1
 $SERVICEUP = 0
 Do {
-    $result = (curl -v http://localhost:8080/auth/).StatusCode
+    $result = (curl -UseBasicParsing -v http://localhost:8080/auth/).StatusCode
     if ($result -eq "200") {
         $SERVICEUP = 1
     } else {
@@ -30,9 +30,9 @@ if ($SERVICEUP -ne 1) {
     throw "Identity Service timed out "
 }
 
-checkStatus (curl -v http://localhost:8080/auth/).StatusCode "200"
-checkStatus (curl -v http://localhost:8080/auth/admin/alfresco/console/).StatusCode "200"
-checkStatus (curl -sSL "http://localhost:8080/auth/realms/alfresco/account").Content "Identity"
+checkStatus (curl -UseBasicParsing -v http://localhost:8080/auth/).StatusCode "200"
+checkStatus (curl -UseBasicParsing -v http://localhost:8080/auth/admin/alfresco/console/).StatusCode "200"
+checkStatus (curl -UseBasicParsing -v "http://localhost:8080/auth/realms/alfresco/protocol/openid-connect/auth?client_id=security-admin-console&redirect_uri=http%3A%2F%2Flocalhost%3A8080%2Fauth%2Fadmin%2Falfresco%2Fconsole%2F&state=ea46ea9f-c963-4f06-89a9-ea9ec04c9694&response_mode=fragment&response_type=code&scope=openid&nonce=44e4af22-2f82-47ef-864e-364c662ae884&code_challenge=IfHxRz3ftCUq4h-SXrSsfXGhoH5z-NfkUSCVyPiNEIc&code_challenge_method=S256").Content "Alfresco Identity Service"
 $Body = @{
    client_id = "alfresco"
    username = "admin"