+++ test/saml/src/test/java/org/alfresco/identity/service/saml/test/TokenTest.java
@@ -8,31 +8,20 @@ import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.lang.reflect.Type;
-import java.net.MalformedURLException;
 import java.security.KeyFactory;
 import java.security.KeyManagementException;
 import java.security.KeyStoreException;
 import java.security.NoSuchAlgorithmException;
 import java.security.interfaces.RSAPublicKey;
-import java.security.spec.InvalidKeySpecException;
 import java.security.spec.X509EncodedKeySpec;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Objects;
 import java.util.Properties;
+import java.util.concurrent.TimeUnit;
 
 import javax.net.ssl.SSLContext;
 
-import com.auth0.jwt.JWT;
-import com.auth0.jwt.JWTVerifier;
-import com.auth0.jwt.algorithms.Algorithm;
-import com.auth0.jwt.exceptions.JWTVerificationException;
-import com.auth0.jwt.interfaces.DecodedJWT;
-import com.google.gson.Gson;
-import com.google.gson.reflect.TypeToken;
-
-import org.openqa.selenium.By.ByLinkText;
-import org.openqa.selenium.By.ByName;
-import org.openqa.selenium.chrome.ChromeDriverService;
 import org.apache.commons.codec.binary.Base64;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.http.HttpEntity;
@@ -47,105 +36,158 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
 import org.junit.jupiter.api.TestInstance.Lifecycle;
+import org.openqa.selenium.By.ByName;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
-import org.openqa.selenium.htmlunit.HtmlUnitDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.chrome.ChromeOptions;
+import org.openqa.selenium.remote.CapabilityType;
 import org.openqa.selenium.remote.DesiredCapabilities;
-import org.openqa.selenium.remote.RemoteWebDriver;
-import org.openqa.selenium.support.ui.Select;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.auth0.jwt.JWT;
+import com.auth0.jwt.JWTVerifier;
+import com.auth0.jwt.algorithms.Algorithm;
+import com.auth0.jwt.exceptions.JWTVerificationException;
+import com.auth0.jwt.interfaces.DecodedJWT;
+import com.google.gson.Gson;
+import com.google.gson.reflect.TypeToken;
+import io.github.bonigarcia.wdm.WebDriverManager;
 
 /**
  * @author Jared Ottley
  */
 @TestInstance(Lifecycle.PER_CLASS)
- public class TokenTest
-{  
-    private Logger logger = LoggerFactory.getLogger(TokenTest.class);
+public class TokenTest
+{
+    private final Logger logger = LoggerFactory.getLogger(TokenTest.class);
+
     private Properties appProps = null;
-    private ChromeDriverService service = null;
+    private WebDriver  driver;
 
     @BeforeAll
     void setup()
     {
         try
         {
-        String rootPath = Thread.currentThread().getContextClassLoader().getResource("").getPath();
-        String appConfigPath = rootPath + "application.properties";
- 
-        appProps = new Properties();
-        appProps.load(new FileInputStream(appConfigPath));
+            String rootPath = Objects.requireNonNull(Thread.currentThread()
+                        .getContextClassLoader()
+                        .getResource(""))
+                        .getPath();
+            String appConfigPath = rootPath + "application.properties";
+
+            appProps = new Properties();
+            appProps.load(new FileInputStream(appConfigPath));
         }
         catch (IOException exception)
         {
             logger.info("Unable to read properties file");
         }
+        driver = createWebDriver();
+    }
+
+    public WebDriver createWebDriver()
+    {
+        return createChromeWebDriver();
+    }
+
+    private WebDriver createChromeWebDriver()
+    {
+            WebDriverManager.chromedriver()
+                        .setup();
+        return new ChromeDriver(getDesiredCapabilities());
+    }
+
+    private DesiredCapabilities getDesiredCapabilities()
+    {
+        DesiredCapabilities capabilities = DesiredCapabilities.chrome();
+        capabilities.setCapability(CapabilityType.ACCEPT_SSL_CERTS, true);
+
+        ChromeOptions chromeOptions = new ChromeOptions();
+        chromeOptions.addArguments("--no-sandbox");
+        chromeOptions.addArguments("--disable-gpu");
+        chromeOptions.addArguments("--disable-dev-shm-usage");
+        chromeOptions.addArguments("--disable-extensions");
+        chromeOptions.addArguments("--single-process");
+        chromeOptions.addArguments("--headless");
+        chromeOptions.addArguments("--test-type");
+        chromeOptions.addArguments("--start-maximized");
+        chromeOptions.addArguments("--disable-web-security");
+        chromeOptions.addArguments("--allow-running-insecure-content");
+        chromeOptions.addArguments("--disable-popup-blocking");
+        chromeOptions.addArguments("--allow-insecure-localhost");
+        chromeOptions.addArguments("--ignore-ssl-errors=yes");
+        chromeOptions.addArguments("--ignore-certificate-errors");
+        chromeOptions.setExperimentalOption("excludeSwitches", new String[] { "enable-automation" });
+        // chromeOptions.addArguments(String.format("--lang=%s", getBrowserLanguage(properties)));
+        //disable profile password manager
+        HashMap<String, Object> chromePrefs = new HashMap<>();
+        chromePrefs.put("credentials_enable_service", false);
+        chromePrefs.put("profile.password_manager_enabled", false);
+        // chromePrefs.put("download.default_directory", getDownloadLocation());
+        chromeOptions.setExperimentalOption("prefs", chromePrefs);
+
+        capabilities.setCapability(ChromeOptions.CAPABILITY, chromeOptions);
+        return capabilities;
     }
 
     @AfterAll
     void cleanup()
     {
         //Clean up remote chromedriver service
-        if (service != null)
+        if (driver != null)
         {
-            service.stop();
+            driver.quit();
         }
     }
 
     @Test
-    void getTokenFromSAMLLogin() 
-    throws MalformedURLException, 
-           IOException, 
-           NoSuchAlgorithmException, 
-           InvalidKeySpecException, 
-           Exception
-    {   
-        //Create HTMLUnit WebDriver
-        WebDriver driver;
+    void getTokenFromSAMLLogin() throws Exception
+    {
+        //Increase Default Timeout for pages to load
+        driver.manage()
+                    .timeouts()
+                    .implicitlyWait(10, TimeUnit.SECONDS);
 
-        if(!isBrowserEnable())
-        {
-            driver = new HtmlUnitDriver(true);
-        }
-        else
-        {
-            ChromeDriverService service = new ChromeDriverService.Builder().usingAnyFreePort().build();
-            service.start();
-            driver = new RemoteWebDriver(service.getUrl(), DesiredCapabilities.chrome());
-        }
-        
         //Initiate page
-        driver.get("https://" + getHostname() + "/auth/realms/" + getRealm() +"/protocol/openid-connect/auth?response_type=id_token%20token&client_id=alfresco&state=CIteJYtFrA22JnCikKHJ2QPrNuGHzyOphE1SsSNs&redirect_uri=http%3A%2F%2F" + getHostname() + "%2Fdummy_redirect&scope=openid%20profile%20email&nonce=CIteJYtFrA22JnCikKHJ2QPrNuGHzyOphE1SsSNs");
-        logger.info("Login page URL: " + driver.getCurrentUrl());
+        driver.get(getBaseUrl() + "/auth/realms/" + getRealm()
+                               + "/protocol/openid-connect/auth?response_type=id_token%20token&client_id=alfresco&state=CIteJYtFrA22JnCikKHJ2QPrNuGHzyOphE1SsSNs&redirect_uri="
+                               + getBaseUrl()
+                               + "/dummy_redirect&scope=openid%20profile%20email&nonce=CIteJYtFrA22JnCikKHJ2QPrNuGHzyOphE1SsSNs");
 
+        logger.info("Login page URL: " + driver.getCurrentUrl());
         //Click on SAML link on login page, the link is theme-dependent
         String themeName = getTheme();
         WebElement element = driver.findElement(
-            themeName.compareTo(ALFRESCO_THEME_NAME) == 0 ? ELEMENT_SAML_ALFRESCO : ELEMENT_SAML_KEYCLOAK);
+                    themeName.compareTo(ALFRESCO_THEME_NAME) == 0 ? ELEMENT_SAML_ALFRESCO : ELEMENT_SAML_KEYCLOAK);
         element.click();
 
         //Select User, Enter password, and submit form on SAML page
-        Select select = new Select(driver.findElement(ByName.name(TokenTestConstants.ELEMENT_USERID)));
-        select.selectByVisibleText(getUser());
+        WebElement usernameField = driver.findElement(ByName.name(TokenTestConstants.ELEMENT_USERID));
+        usernameField.sendKeys(getUser());
         WebElement passwordField = driver.findElement(ByName.name(TokenTestConstants.ELEMENT_PASWORD));
         passwordField.sendKeys(getPassword());
-
         passwordField.submit();
-        
+
+        // Workaround to get the tests passing when using 'http' rather than 'https' protocol
+        Thread.sleep(3000L);
+
         //Get the redirect URL for validation -- If you check the status of the
         //redirct URL call it will be 404.  The page does not exist. All we are
         //intersted in is the token parameter in the URL
         logger.info("Redirect URL: " + driver.getCurrentUrl());
+        logger.info("Page title: " + driver.getTitle());
 
         //Get token param
         Map<String, String> params = getQueryStringMap(driver.getCurrentUrl());
+        logger.info("URL Params: " + params);
+
         String token = params.get(TokenTestConstants.HEADER_ACCESS_TOKEN);
         logger.info("access_token parameter: " + token);
-        
 
         //Decode token and verify token
+
         DecodedJWT jwt = null;
         try
         {
@@ -161,21 +203,18 @@ import org.slf4j.LoggerFactory;
             Algorithm algorithm = Algorithm.RSA256(pubKey, null);
             logger.info("issuer: " + getIssuer());
             JWTVerifier verifier = JWT.require(algorithm)
-                .withIssuer(getIssuer())
-                .build(); //Reusable verifier instance
+                        .withIssuer(getIssuer())
+                        .build(); //Reusable verifier instance
             jwt = verifier.verify(token);
-            logger.info("Payload Decoded: " + new String(Base64.decodeBase64(jwt.getPayload().getBytes())));
+            logger.info("Payload Decoded: " + new String(Base64.decodeBase64(jwt.getPayload()
+                                                                                         .getBytes())));
         }
         catch (JWTVerificationException exception)
         {
-            logger.info("Verfication failed");
+            logger.info("Verification failed");
         }
 
-        assertNotNull(jwt);  
-        
-        //Quit Driver session
-        driver.quit();
-
+        assertNotNull(jwt);
     }
 
     //Utility Methods
@@ -187,7 +226,7 @@ import org.slf4j.LoggerFactory;
         {
             String[] split = url.split("#");
 
-            if(split.length == 2 && StringUtils.isNotBlank(split[1]))
+            if (split.length == 2 && StringUtils.isNotBlank(split[1]))
             {
                 String[] params = split[1].split("&");
 
@@ -199,7 +238,8 @@ import org.slf4j.LoggerFactory;
 
                         if (working.length >= 1 && StringUtils.isNotBlank(working[0]))
                         {
-                            map.put(working[0], ((working.length == 2 && StringUtils.isNotBlank(working[1])) ? working[1] : ""));
+                            map.put(working[0],
+                                    ((working.length == 2 && StringUtils.isNotBlank(working[1])) ? working[1] : ""));
                         }
                     }
                 }
@@ -216,24 +256,23 @@ import org.slf4j.LoggerFactory;
         try
         {
             //Create HTTP Client that allows any ssl cert
-            SSLContext sslContext = new SSLContextBuilder()
-                .loadTrustMaterial(null, (certificate, authType) -> true).build();
-            
+            SSLContext sslContext = new SSLContextBuilder().loadTrustMaterial(null, (certificate, authType) -> true)
+                        .build();
+
             CloseableHttpClient httpClient = HttpClients.custom()
-                .setSSLContext(sslContext)
-                .setSSLHostnameVerifier(new NoopHostnameVerifier())
-                .build();
+                        .setSSLContext(sslContext)
+                        .setSSLHostnameVerifier(new NoopHostnameVerifier())
+                        .build();
 
             //Get realm details
-            HttpGet httpGet = new HttpGet("https://" + getHostname() + "/auth/realms/" + getRealm() + "/");
+            HttpGet httpGet = new HttpGet(getBaseUrl() + "/auth/realms/" + getRealm() + "/");
 
             CloseableHttpResponse response = httpClient.execute(httpGet);
             HttpEntity entity = response.getEntity();
-            
-            BufferedReader in = new BufferedReader(
-            new InputStreamReader(entity.getContent()));
+
+            BufferedReader in = new BufferedReader(new InputStreamReader(entity.getContent()));
             String inputLine;
-            StringBuffer content = new StringBuffer();
+            StringBuilder content = new StringBuilder();
             while ((inputLine = in.readLine()) != null)
             {
                 content.append(inputLine);
@@ -242,9 +281,11 @@ import org.slf4j.LoggerFactory;
 
             //close HTTP Client
             httpClient.close();
-            
+
             //Parse Response
-            Type type = new TypeToken<Map<String, String>>(){}.getType();
+            Type type = new TypeToken<Map<String, String>>()
+            {
+            }.getType();
             Gson gson = new Gson();
             Map<String, String> responseMap = gson.fromJson(content.toString(), type);
 
@@ -261,25 +302,40 @@ import org.slf4j.LoggerFactory;
         return key;
     }
 
+    private String getProtocol()
+    {
+        return resolveProperty(PROP_KEYCLOAK_PROTOCOL);
+    }
+
     private String getHostname()
     {
         return resolveProperty(PROP_KEYCLOAK_HOSTNAME);
     }
 
+    private String getPort()
+    {
+        String port = resolveProperty(PROP_KEYCLOAK_PORT);
+        return Integer.parseInt(port) == 80 ? "" : ":" + port;
+    }
+
+    private String getBaseUrl()
+    {
+        String url = getProtocol() + "://" + getHostname() + getPort();
+        logger.info("Base URL: " + url);
+        return url;
+    }
+
     /**
      * Return who we expect to have created and sigend the token.
-     * 
-     * If not provied in the the environment or properties file we will attempt to build it from the hostname and realm
-     * 
-     * @return
+     * If not provided in the the environment or properties file we will attempt to build it from the hostname and realm
      */
     private String getIssuer()
     {
         String keycloak_issuer = resolveProperty(PROP_KEYCLOAK_ISSUER);
-        
+
         if (StringUtils.isEmpty(keycloak_issuer))
         {
-            keycloak_issuer = buildIssuer(getHostname(), getRealm());
+            keycloak_issuer = buildIssuer();
         }
 
         return keycloak_issuer;
@@ -292,7 +348,7 @@ import org.slf4j.LoggerFactory;
 
     private String getUser()
     {
-        return resolveProperty(PROP_SAML_USERNAME);
+        return resolveProperty(SAML_USERNAME);
     }
 
     private String getTheme()
@@ -305,23 +361,11 @@ import org.slf4j.LoggerFactory;
         return resolveProperty(PROP_SAML_PASSWORD);
     }
 
-    private Boolean isBrowserEnable()
-    {
-        String enabled = resolveProperty(PROP_ENABLE_BROWSER);
-
-        if (StringUtils.isNotBlank(enabled))
-        {
-            return Boolean.valueOf(enabled);
-        }
-
-        return false;
-    }
-
     /**
      * Return property value
      * <BR>
-     * Values can be overridden from environment variables. In this case the following naming convention applies:
-     * <P>
+     * Values can be overridden from environment variables and Java properties. In this case the following naming convention applies:
+     * <p>
      * <code>property.name (property) -> PROPERTY_NAME (environment variable)</code>
      *
      * @param propertyName property name
@@ -329,25 +373,29 @@ import org.slf4j.LoggerFactory;
      */
     private String resolveProperty(String propertyName)
     {
-        String envVarName = propertyName.replace(".", "_").toUpperCase();
+        String envVarName = propertyName.replace(".", "_")
+                    .toUpperCase();
         String propertyValue = System.getenv(envVarName);
-
         if (StringUtils.isEmpty(propertyValue))
         {
-            propertyValue = appProps.getProperty(propertyName);
+            // Check Java properties too
+            propertyValue = System.getProperty(propertyName);
+
+            if (StringUtils.isEmpty(propertyValue))
+            {
+                propertyValue = appProps.getProperty(propertyName);
+            }
         }
 
         return propertyValue;
     }
 
-    private String buildIssuer(String hostname, String realm)
+    private String buildIssuer()
     {
-        if (StringUtils.isNotEmpty(hostname) && StringUtils.isNotEmpty(realm))
+        if (StringUtils.isNotEmpty(getHostname()) && StringUtils.isNotEmpty(getRealm()))
         {
-            return "https://" + hostname + "/auth/realms/" + realm;
+            return getBaseUrl() + "/auth/realms/" + getRealm();
         }
-
         return "";
     }
-    
 }