+++ test/saml/src/test/java/org/alfresco/identity/service/saml/test/TokenTestConstants.java
@@ -9,20 +9,22 @@ import org.openqa.selenium.By;
 public final class TokenTestConstants
 {
     //Property File Keys
-    public final static String PROP_SAML_USERNAME = "saml.username";
-    public final static String PROP_SAML_PASSWORD = "saml.password";
+    public final static String SAML_USERNAME          = "saml.username";
+    public final static String PROP_SAML_PASSWORD     = "saml.password";
+    public static final String PROP_KEYCLOAK_PROTOCOL = "keycloak.protocol";
     public final static String PROP_KEYCLOAK_HOSTNAME = "keycloak.hostname";
-    public final static String PROP_KEYCLOAK_ISSUER = "keycloak.issuer";
-    public final static String PROP_KEYCLOAK_REALM = "keycloak.realm";
-    public final static String PROP_KEYCLOAK_THEME = "keycloak.theme";
-    public final static String PROP_ENABLE_BROWSER = "enable.browser";
+    public static final String PROP_KEYCLOAK_PORT     = "keycloak.port";
+    public final static String PROP_KEYCLOAK_ISSUER   = "keycloak.issuer";
+    public final static String PROP_KEYCLOAK_REALM    = "keycloak.realm";
+    public final static String PROP_KEYCLOAK_THEME    = "keycloak.theme";
+    public final static String PROP_ENABLE_BROWSER    = "enable.browser";
 
     //Page Elements
     // SAML button selector varies depending on applied theme
     public final static By ELEMENT_SAML_KEYCLOAK = By.linkText("saml");
     public final static By ELEMENT_SAML_ALFRESCO = By.cssSelector("input[value=saml]");
 
-    public final static String ELEMENT_USERID = "userid";
+    public final static String ELEMENT_USERID = "username";
     public final static String ELEMENT_PASWORD = "password";
 
     //Header Parameters